Update network_map.html
Can you make a few changes and give me an updated file? - In the legend, add a total number of networks on the right after "Networks by County" phrase. Make it the same style as the numbers by each county. - Add a link to each of the popups after "Email" called "Churches" and link it to /org_map and pass a parameter of "network" with the network name - make the text of the network name that shows up over the polygons responsive so that it's smaller when the map is shown on a mobile device or small screen and larger when shown on the desktop