Update org_map.html
I updated your org_map.html so that the Category and Denomination filters are hidden if their tag lists are empty. I did this by checking the length of the tag arrays before building the lists. If the array is empty, the corresponding <details> block (#grp-category or #grp-denomination) is hidden with style.display = "none"